Multiple improvements:
Added and used directives for profile, channel and chat pages.
Cleanup of controllers from code that was moved to the directives.
Socket messages in app.run() use $scope.$apply() always to avoid using the same in the controllers.
Created a Popup service and PopupCtrl used to show popups from everywhere with dynamically provided title and text.
Socket.disconnect is shown as error in the console's log.
Moved joinUrl logic from the ChatCtrl to the Chat service.